🚚 Move: `tag-picker/*.svelte` to `$lib/components/tag/picker/` [#90]
